packages/telegram-bot/pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-bot"
version = "0.1.0"
description = "Telegram bot interface for the Job Agent platform"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.2.1",
    "job-agent-backend >=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "black>=25.9.0",
    "ruff>=0.14.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
